All functional logic done, just needs unit tests
