+++ src/components/common/common-components.module.ts
@@ -29,6 +29,7 @@ import { CPCDebriefCardComponent } from './cpc-debrief-card/cpc-debrief-card';
 import { PipesModule } from '../../shared/pipes/pipes.module';
 import { HealthDeclarationSignedComponent } from './health-declaration-signed/health-declaration-signed';
 import { BikeCategoryTypeComponent } from './bike-category-type/bike-category-type';
+import {TransmissionDisplayComponent} from './transmission-display/transmission-display';
 
 @NgModule({
   declarations: [
@@ -56,6 +57,7 @@ import { BikeCategoryTypeComponent } from './bike-category-type/bike-category-ty
     CPCDebriefCardComponent,
     HealthDeclarationSignedComponent,
     BikeCategoryTypeComponent,
+    TransmissionDisplayComponent,
   ],
   imports: [
     SignaturePadModule,
@@ -90,6 +92,7 @@ import { BikeCategoryTypeComponent } from './bike-category-type/bike-category-ty
     PipesModule,
     HealthDeclarationSignedComponent,
     BikeCategoryTypeComponent,
+    TransmissionDisplayComponent,
   ],
 })
 export class ComponentsModule { }
