+++ src/components/common/transmission-display/transmission-display.html
@@ -1,0 +1,4 @@
+<data-row
+  [label]="'Transmission'"
+  [value]="getTransmissionText(transmission, code78, category)">
+</data-row>
