+++ src/pages/view-test-result/cat-d/view-test-result.cat-d.page.html
@@ -25,7 +25,11 @@
       <rekey-reason-card [data]='testResult.rekeyReason'></rekey-reason-card>
     </div>
     <examiner-details-card [data]='getExaminerDetails()'></examiner-details-card>
-    <vehicle-details-card [data]='testResult.vehicleDetails'></vehicle-details-card>
+<!--    <vehicle-details-card [data]='testResult.vehicleDetails'></vehicle-details-card>-->
+    <vehicle-details-card
+      [data]='testResult.vehicleDetails'
+      [category]="testResult.category"
+      [passCompletion]='testResult.passCompletion'></vehicle-details-card>
     <debrief-card [data]='testResult.testData' [category]="testResult.category"></debrief-card>
     <test-summary-card
       [accompaniment]='testResult.accompaniment'
